Restore tests like on develop
