+++ tests-legacy/Endpoints/AbstractEndpointAdminTest.php
@@ -32,20 +32,10 @@ use Context;
 use Employee;
 use PhpEncryption;
 use PrestaShop\PrestaShop\Adapter\SymfonyContainer;
-use PrestaShop\PrestaShop\Core\Domain\Employee\AuthorizationOptions;
-use PrestaShopBundle\Security\Admin\Employee as LoggedEmployee;
 use ReflectionClass;
-use Symfony\Component\Security\Core\Authentication\Token\AbstractToken;
-use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorage;
-use Symfony\Component\Security\Core\Role\Role;
 
 abstract class AbstractEndpointAdminTest extends AbstractEndpointTest
 {
-    /**
-     * @var TokenStorage
-     */
-    private $tokenStorage;
-
     protected function setUp()
     {
         parent::setUp();
@@ -54,15 +44,6 @@ abstract class AbstractEndpointAdminTest extends AbstractEndpointTest
             define('_PS_TAB_MODULE_LIST_URL_', '');
         }
         Context::getContext()->employee = new Employee(1);
-
-        $this->tokenStorage = SymfonyContainer::getInstance()->get('security.token_storage');
-    }
-
-    protected function tearDown()
-    {
-        parent::tearDown();
-
-        SymfonyContainer::getInstance()->set('security.token_storage', $this->tokenStorage);
     }
 
     protected function employeeLogin()
@@ -73,54 +54,6 @@ abstract class AbstractEndpointAdminTest extends AbstractEndpointTest
         $cookieName = 'PrestaShop-' . md5(_PS_VERSION_ . 'psAdmin');
         $_COOKIE[$cookieName] = $cipherTool->encrypt($cookieContent);
         Cache::store('isLoggedBack' . 1, true);
-
-        $this->symfonyLogIn();
-    }
-
-    /**
-     * Emulates a real employee logged to the Back Office.
-     */
-    protected function symfonyLogIn()
-    {
-        $loggedEmployeeData = new \stdClass();
-        $loggedEmployeeData->email = 'demo@prestashop.com';
-        $loggedEmployeeData->id = 1;
-        $loggedEmployeeData->passwd = '';
-        $loggedEmployeeMock = new LoggedEmployee($loggedEmployeeData);
-
-        $tokenMock = $this
-            ->getMockBuilder(AbstractToken::class)
-            ->setMethods([
-                'getUser',
-                'getRoles',
-                'isAuthenticated',
-            ])
-            ->getMockForAbstractClass();
-
-        $tokenMock->expects($this->any())
-            ->method('getUser')
-            ->willReturn($loggedEmployeeMock);
-
-        $tokenMock->expects($this->any())
-            ->method('getRoles')
-            ->willReturn([new Role(AuthorizationOptions::DEFAULT_EMPLOYEE_ROLE)]);
-
-        $tokenMock->expects($this->any())
-            ->method('isAuthenticated')
-            ->willReturn(true);
-
-        $tokenStorageMock = $this->getMockBuilder(TokenStorage::class)
-            ->setMethods([
-                'getToken',
-            ])
-            ->disableAutoload()
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $tokenStorageMock->method('getToken')
-            ->willReturn($tokenMock);
-
-        SymfonyContainer::getInstance()->set('security.token_storage', $tokenStorageMock);
     }
 
     /**
