+++ tests-legacy/Integration/PrestaShopBundle/Controller/Admin/SurvivalTest.php
@@ -27,6 +27,9 @@
 namespace LegacyTests\Integration\PrestaShopBundle\Controller\Admin;
 
 use LegacyTests\Integration\PrestaShopBundle\Test\WebTestCase;
+use PrestaShopBundle\Security\Admin\Employee as LoggedEmployee;
+use Symfony\Component\Security\Core\Authentication\Token\AbstractToken;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorage;
 
 /**
  * @group demo
@@ -36,11 +39,34 @@ use LegacyTests\Integration\PrestaShopBundle\Test\WebTestCase;
 class SurvivalTest extends WebTestCase
 {
     /**
+     * @var TokenStorage
+     */
+    private $tokenStorage;
+
+    /**
      * {@inheritdoc}
      */
     public static function setUpBeforeClass()
     {
         // Do not reset the Database.
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function setUp()
+    {
+        parent::setUp();
+        $this->tokenStorage = self::$kernel->getContainer()->get('security.token_storage');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function tearDown()
+    {
+        self::$kernel->getContainer()->set('security.token_storage', $this->tokenStorage);
+        parent::tearDown();
     }
 
     /**
@@ -148,8 +174,53 @@ class SurvivalTest extends WebTestCase
             'admin_themes_import' => ['Add new theme', 'admin_themes_import'],
             'admin_mail_theme_index' => ['Email Theme', 'admin_mail_theme_index'],
             'admin_emails_index' => ['Advanced parameters E-mail', 'admin_emails_index'],
-            '_admin_login' => ['Login', '_admin_login'],
-            '_admin_forgot_password' => ['Forgot password (login)', '_admin_forgot_password'],
         ];
+    }
+
+    /**
+     * Emulates a real employee logged to the Back Office.
+     * For survival tests only.
+     */
+    private function logIn()
+    {
+        $loggedEmployeeData = new \stdClass();
+        $loggedEmployeeData->email = 'demo@prestashop.com';
+        $loggedEmployeeData->id = 1;
+        $loggedEmployeeData->passwd = '';
+        $loggedEmployeeMock = new LoggedEmployee($loggedEmployeeData);
+
+        $tokenMock = $this
+            ->getMockBuilder(AbstractToken::class)
+            ->setMethods([
+                'getUser',
+                'getRoles',
+                'isAuthenticated',
+            ])
+            ->getMockForAbstractClass();
+
+        $tokenMock->expects($this->any())
+            ->method('getUser')
+            ->willReturn($loggedEmployeeMock);
+
+        $tokenMock->expects($this->any())
+            ->method('getRoles')
+            ->willReturn([]);
+
+        $tokenMock->expects($this->any())
+            ->method('isAuthenticated')
+            ->willReturn(true);
+
+        $tokenStorageMock = $this->getMockBuilder(TokenStorage::class)
+            ->setMethods([
+                'getToken',
+            ])
+            ->disableAutoload()
+            ->disableOriginalConstructor()
+            ->getMock();
+
+        $tokenStorageMock->method('getToken')
+            ->willReturn($tokenMock);
+
+        self::$kernel->getContainer()->set('security.token_storage', $tokenStorageMock);
     }
 }
