+++ tests-legacy/Integration/PrestaShopBundle/Controller/Api/ApiTestCase.php
@@ -28,7 +28,6 @@ namespace LegacyTests\Integration\PrestaShopBundle\Controller\Api;
 
 use Context;
 use Language;
-use LegacyTests\Services\EmployeeLoginTrait;
 use PrestaShop\PrestaShop\Adapter\LegacyContext;
 use Shop;
 use PrestaShop\PrestaShop\Core\Addon\Theme\Theme;
@@ -38,8 +37,6 @@ use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 // bin/phpunit -c tests/phpunit-admin.xml --group api --stop-on-error --stop-on-failure --verbose --debug
 abstract class ApiTestCase extends WebTestCase
 {
-    use EmployeeLoginTrait;
-
     /**
      * @var \Symfony\Component\Routing\RouterInterface
      */
@@ -70,8 +67,11 @@ abstract class ApiTestCase extends WebTestCase
         $this->oldContext = Context::getContext();
         $legacyContextMock = $this->mockContextAdapter();
         self::$container->set('prestashop.adapter.legacy.context', $legacyContextMock);
-        self::$client = $this->createClient();
-        $this->logIn();
+
+        $client = self::$kernel->getContainer()->get('test.client');
+        $client->setServerParameters(array());
+
+        self::$client = $client;
     }
 
     protected function tearDown()
@@ -146,7 +146,6 @@ abstract class ApiTestCase extends WebTestCase
     {
         $employeeMock = $this->getMockBuilder('\Employee')->getMock();
         $employeeMock->id_lang = 1;
-        $employeeMock->id_profile = _PS_ADMIN_PROFILE_;
 
         return $employeeMock;
     }
