+++ tests-legacy/Integration/PrestaShopBundle/Controller/Api/TranslationControllerTest.php
@@ -29,7 +29,6 @@ namespace LegacyTests\Integration\PrestaShopBundle\Controller\Api;
 /**
  * @group api
  * @group translation
- * @runTestsInSeparateProcesses
  */
 class TranslationControllerTest extends ApiTestCase
 {
