+++ tests-legacy/PrestaShopBundle/Controller/ControllerTest.php
@@ -117,6 +117,7 @@ class ControllerTest extends TestCase
             array('AdminCarriersController'),
             array('AdminStatusesController'),
             array('AdminZonesController'),
+            array('AdminLoginController'),
             array('AdminQuickAccessesController'),
             array('AdminCustomerThreadsController'),
             array('AdminReferrersController'),
