+++ tests-legacy/phpunit-admin.xml
@@ -1,6 +1,6 @@
 <phpunit
   bootstrap="bootstrap-admin.php"
-  processIsolation="true"
+  processIsolation="false"
   convertErrorsToExceptions="true"
   convertNoticesToExceptions="true"
   convertWarningsToExceptions="true"
@@ -9,7 +9,6 @@
         <env name="SYMFONY_ENV" value="test"/>
         <env name="kernel.environment" value="test" />
         <server name="KERNEL_CLASS" value="AppKernel" />
-        <server name="REMOTE_ADDR" value=""/>
         <const name="_PS_IN_TEST_" value="1" />
     </php>
     <listeners>
